+++ bindings/matrix-sdk-ffi/src/qr_code.rs
@@ -21,7 +21,7 @@ use matrix_sdk::authentication::oauth::{
     },
     OAuth,
 };
-use matrix_sdk_base::crypto::types::qr_login;
+use matrix_sdk_base::crypto::types::qr_login::{self, QrCodeIntent};
 use matrix_sdk_common::{stream::StreamExt, SendOutsideWasm, SyncOutsideWasm};
 
 use crate::{
@@ -272,6 +272,26 @@ impl QrCodeData {
             },
             qr_login::QrCodeIntentData::Msc4388 { .. } => None,
         }
+    }
+
+    /// The base URL of the homeserver contained within the scanned QR code
+    /// data.
+    ///
+    /// Note: This value is only present when scanning a QR code conforming to
+    /// MSC4388.
+    pub fn base_url(&self) -> Option<String> {
+        match self.inner.intent_data() {
+            qrcode::QrCodeIntentData::Msc4108 { .. } => None,
+            qrcode::QrCodeIntentData::Msc4388 { base_url, .. } => Some(base_url.to_string()),
+        }
+    }
+
+    /// Get the [`QrCodeIntent`] of this [`QrCodeData`] object.
+    ///
+    /// This tells us if the creator of the QR code wants to log in or if they
+    /// want to log another device in.
+    pub fn intent(&self) -> QrCodeIntent {
+        self.inner.intent()
     }
 }